Add goal increment

+++ PinguinEducation/PinguinEducation/Alphabet/AlphabetStorage.cs
@@ -10,6 +10,6 @@ namespace PinguinEducation.Alphabet
             { return new string[] { "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z" }; } }
         public string ALPHABET_letter { get; set; }
         public string[] ALPHABET_unlocked { get; set; }
-        public int goal { get; set; }
+        public double goal { get; set; }
     }
 }
